+++ bs-config.json
@@ -1,0 +1,8 @@
+{
+  "server": {
+    "baseDir": "./public",
+    "routes": {
+      "/dist": "dist"
+    }
+  }
+}
