--- index.html
@@ -1,10 +1,0 @@
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Document</title>
-    <script src="geofence.min.js" ></script>
-</head>
-<body>
-  <div id="map"></div>
-  <div id="info"></div>
-</body>
