+++ public/index.html
@@ -1,0 +1,22 @@
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>Document</title>
+  <link rel="stylesheet" href="style.css">
+  <script type="module" src="dist/index.js"></script>
+</head>
+
+<body>
+  <h2>Reqired Location</h2>
+  <div id="required-location-map-outer">
+    <div id="required-location-map"></div>
+  </div>
+  <h2>Your Location</h2>
+  <div id="your-location-map"></div>
+  <h2>Info</h2>
+  <div id="info">
+    <button id="register-button">Register</button>
+  </div>
+</body>
+</html>
